put arena layouts into different file
