+++ BossMod/Modules/MaskedCarnivale/Stage07AChorusSlime/Stage07Act3.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using System.Linq;
 using BossMod.Components;
 
@@ -55,35 +54,12 @@ class Stage07Act3States : StateMachineBuilder
         }
     }
 
-public class Stage07Act3 : BossModule
+public class Stage07Act3 : Layout2Corners
     {
-        public static IEnumerable<WPos> Wall1()
-        {
-            yield return new WPos(85,95);
-            yield return new WPos(95,95);
-            yield return new WPos(95,89);
-            yield return new WPos(94.5f,89);
-            yield return new WPos(94.5f,94.5f);
-            yield return new WPos(85,94.5f);
-        }
-        public static IEnumerable<WPos> Wall2()
-        {
-            yield return new WPos(105,95);
-            yield return new WPos(115,95);
-            yield return new WPos(115,94.5f);
-            yield return new WPos(105.5f,94.5f);
-            yield return new WPos(105.5f,89);
-            yield return new WPos(105,89);
-        }
         public Stage07Act3(WorldState ws, Actor primary) : base(ws, primary, new ArenaBoundsCircle(new(100, 100), 25))
         {
             ActivateComponent<Hints>();
             ActivateComponent<SlimeExplosion>();
-        }
-        protected override void DrawArenaForeground(int pcSlot, Actor pc)
-        {
-            Arena.AddPolygon(Wall1(),ArenaColor.Border);
-            Arena.AddPolygon(Wall2(),ArenaColor.Border);
         }
         protected override bool CheckPull() { return PrimaryActor.IsTargetable && PrimaryActor.InCombat || Enemies(OID.Slime).Any(e => e.InCombat); }
         protected override void DrawEnemies(int pcSlot, Actor pc)
