+++ BossMod/Modules/MaskedCarnivale/Stage13BeautyandaBeast/Stage13Act2.cs
@@ -8,24 +8,24 @@ namespace BossMod.MaskedCarnivale.Stage13.Act2
         Succubus = 0x26F7, //R=1.0
         Helper = 0x233C, //R=0.5
     };
-public enum AID : uint
-{
-    Attack = 6497, // 26F8/26F7->player, no cast, single-target
-    VoidFireII = 14880, // 26F8->location, 3,0s cast, range 5 circle
-    VoidAero = 14881, // 26F8->self, 3,0s cast, range 40+R width 8 rect
-    DarkSabbath = 14951, // 26F8->self, 3,0s cast, range 60 circle, gaze
-    DarkMist = 14884, // 26F8->self, 3,0s cast, range 8+R circle
-    CircleOfBlood = 15043, // 26F8->self, 3,0s cast, single-target
-    CircleOfBlood2 = 14887, // 233C->self, 3,0s cast, range 10-20 donut
-    VoidFireIV = 14888, // 26F8->location, 4,0s cast, range 10 circle
-    VoidFireIV2 = 14886, // 26F8->self, no cast, single-target
-    VoidFireIV3 = 14889, // 233C->location, 3,0s cast, range 6 circle
-    SummonDarkness = 14885, // 26F8->self, no cast, single-target, summons succubus add
-    BeguilingMist = 15045, // 26F7->self, 7,0s cast, range 50+R circle, interruptable, applies hysteria
-    FatalAllure = 14952, // 26F8->self, no cast, range 50+R circle, attract, applies terror
-    BloodRain = 14882, // 26F8->location, 3,0s cast, range 50 circle
+    public enum AID : uint
+    {
+        Attack = 6497, // 26F8/26F7->player, no cast, single-target
+        VoidFireII = 14880, // 26F8->location, 3,0s cast, range 5 circle
+        VoidAero = 14881, // 26F8->self, 3,0s cast, range 40+R width 8 rect
+        DarkSabbath = 14951, // 26F8->self, 3,0s cast, range 60 circle, gaze
+        DarkMist = 14884, // 26F8->self, 3,0s cast, range 8+R circle
+        CircleOfBlood = 15043, // 26F8->self, 3,0s cast, single-target
+        CircleOfBlood2 = 14887, // 233C->self, 3,0s cast, range 10-20 donut
+        VoidFireIV = 14888, // 26F8->location, 4,0s cast, range 10 circle
+        VoidFireIV2 = 14886, // 26F8->self, no cast, single-target
+        VoidFireIV3 = 14889, // 233C->location, 3,0s cast, range 6 circle
+        SummonDarkness = 14885, // 26F8->self, no cast, single-target, summons succubus add
+        BeguilingMist = 15045, // 26F7->self, 7,0s cast, range 50+R circle, interruptable, applies hysteria
+        FatalAllure = 14952, // 26F8->self, no cast, range 50+R circle, attract, applies terror
+        BloodRain = 14882, // 26F8->location, 3,0s cast, range 50 circle
 
-};
+    };
 
     class VoidFireII : LocationTargetedAOEs
     {
