+++ BossMod/Modules/MaskedCarnivale/Stage14BlobsintheWoods/Stage14Act1.cs
@@ -40,34 +40,11 @@ namespace BossMod.MaskedCarnivale.Stage14.Act1
         }
     }
 
-    public class Stage14Act1 : BossModule
+    public class Stage14Act1 : Layout2Corners
     {
-        public static IEnumerable<WPos> Wall1()
-        {
-            yield return new WPos(85,95);
-            yield return new WPos(95,95);
-            yield return new WPos(95,89);
-            yield return new WPos(94.5f,89);
-            yield return new WPos(94.5f,94.5f);
-            yield return new WPos(85,94.5f);
-        }
-        public static IEnumerable<WPos> Wall2()
-        {
-            yield return new WPos(105,95);
-            yield return new WPos(115,95);
-            yield return new WPos(115,94.5f);
-            yield return new WPos(105.5f,94.5f);
-            yield return new WPos(105.5f,89);
-            yield return new WPos(105,89);
-        }
         public Stage14Act1(WorldState ws, Actor primary) : base(ws, primary, new ArenaBoundsCircle(new(100, 100), 25))
         {
             ActivateComponent<Hints>();
-        }
-        protected override void DrawArenaForeground(int pcSlot, Actor pc)
-        {
-            Arena.AddPolygon(Wall1(),ArenaColor.Border);
-            Arena.AddPolygon(Wall2(),ArenaColor.Border);
         }
         protected override void DrawEnemies(int pcSlot, Actor pc)
         {
